==> 01_scape_restaurant_links.ipynb <==
# Extracting Links of Restaurants to Scrape
Notebook used to extract the TripAdvisor links of the restaurants to be scrapped for our study

## 1. Basic Setup

### 1.1 Loading of Libaries

from selenium import webdriver      # conda install -c conda-forge selenium
from time import sleep
from selenium.webdriver.chrome.options import Options

import pandas as pd

import warnings
warnings.filterwarnings("ignore")

### 1.2 Initialising Chromedriver

# configuration tool that can be passed into webdriver.Chrome
options = Options()

# windows
# https://www.youtube.com/watch?v=Xjv1sY630Uc 
# PATH = "C:\Program Files\chromedriver.exe"
# driver = webdriver.Chrome(PATH)

# macOS
# https://www.edureka.co/community/52315/how-to-setup-chrome-driver-with-selenium-on-macos
# need to put chromedriver.exe into /usr/local/bin
# and type in terminal (after cd into /usr/local/bin)
#  xattr -d com.apple.quarantine chromedriver 
driver = webdriver.Chrome()

## 2. Scraping of Links

### 2.1 Accessing website to scrape

# website to scrape
driver.get("https://www.tripadvisor.com.sg/Restaurants-g294265-Singapore.html")     ### Website with full listing of restaurants in Singapore

### 2.2 Extracting the page links of all the restaurants listed

#### 2.2.1 Extracting Links from Page 1

# 'a' tag element that redirects to the specific page of a restaurant
tags = driver.find_elements_by_class_name('bHGqj')

# extract the href value of all these 'a' tag elements
links = [element.get_attribute("href") for element in tags]

#### 2.2.2 Navgiating to subsequent pages

# 'next' button element of pagination that leads to the next page
next_page_button = driver.find_element_by_class_name("next")

while next_page_button.is_displayed():
    try:
        next_page_button.click()    # click on 'next' button
        sleep(5)       # wait for next page items to be fully loaded

        # --- extracting links --- #
        new_tags = driver.find_elements_by_class_name('bHGqj')
        new_links = [element.get_attribute("href") for element in new_tags]
        links.extend(new_links)

        # get 'next' button of current page
        next_page_button = driver.find_element_by_class_name("next")
    except:
        break

### 2.3 Exporting the Links into an Excel file

df_output = pd.DataFrame({'links': links})

df_output.drop_duplicates(inplace=True)

len(df_output)

df_output.to_excel("data/tripadvisor_restaurant_links.xlsx", encoding="utf-8-sig", index=False)
