==> 03_scrape_reviews.ipynb <==
# Extract Restaurant Review
Notebook used to extract the TripAdvisor Restaurant Review data

## 1. Basic Setup

### 1.1 Loading of Libaries

from selenium import webdriver      # conda install -c conda-forge selenium
from time import sleep
from selenium.webdriver.chrome.options import Options

import pandas as pd
import numpy as np
import re

import warnings
warnings.filterwarnings("ignore")

## 2. Read in Dataset with Restaurant Informtaion
- To generate the "Review Pages" to scrape

df_restaurant = pd.read_csv("data/full_restaurant_rating_data.csv")
df_restaurant_subset = df_restaurant[["url", "num_of_reviews"]].copy()

df_restaurant_subset

# Functions that generate the list of "review pages" of a restaurant (given its restaurant url page on TripAdvisor)

def get_list_of_reviews_url(url, num_of_reviews):
    lst_of_urls = [url]
    url_front = re.findall(".*\-Reviews", url)[0]
    url_back = re.findall("(?<=Reviews\-).*",  url)[0]

    max_review_page = int(np.ceil(num_of_reviews/ 10) * 10)
    for i in range(10, max_review_page, 10):
        new_url = url_front + f"-or{i}-" + url_back
        lst_of_urls.append(new_url)

    return lst_of_urls

# variable to control which range of index to scrape from df_restaurant
index_start = 2000;
index_end = 2003;

driver = webdriver.Chrome()

dct_review_links = {}   # key: restaurant url, value: list of review pages url

for idx, row in df_restaurant.loc[index_start:index_end,:].iterrows():
        list_of_urls = get_list_of_reviews_url(row["url"], row["num_of_reviews"])
        counter = len(list_of_urls)
        for i in range(len(list_of_urls)-1, -1, -1):
            driver.get(list_of_urls[i])
            curr_url = driver.current_url
            if curr_url == list_of_urls[0]:
                counter = counter - 1
            else:
                break
        dct_review_links[idx] = [row["url"]]
        print(f"Done {idx} of {index_end}")

df_links = pd.DataFrame()

for key, values in dct_review_links.items():
    df_links = df_links.append(pd.DataFrame({"restaurant_url":[df_restaurant.loc[key, "url"]] * len(values), "review_url": values}))

df_links.reset_index(drop=True, inplace=True)

df_links

## 3. Scraping of Restaurant Reviews

# Function to scrape review information from a page
def scrape_page_review(df, restaurant_url):
    review_container = driver.find_elements_by_class_name("review-container")
    date_of_visit_list = [e.find_element_by_class_name("prw_reviews_stay_date_hsx").text for e in review_container]
    for i in range(0, len(review_container)):
        # ----------------------------- Review Title ----------------------------- #
        review_title = review_container[i].find_element_by_class_name("quote").text

        # ----------------------------- Review Date ----------------------------- #
        review_date = review_container[i].find_element_by_class_name("ratingDate").get_attribute("title")

        # ----------------------------- Review Rating (out of 5) ----------------------------- #
        try:
            review_rating = review_container[i].find_element_by_class_name("ui_bubble_rating").get_attribute("class")
        except:
            return df

        # ----------------------------- Review Description ----------------------------- #
        review_description = review_container[i].find_element_by_class_name("partial_entry").get_attribute("textContent").replace("...More", " ")
        # ----------------------------- More Review Description ----------------------------- #
        try:
            more_description = driver.find_elements_by_class_name("review-container")[i].find_element_by_class_name("postSnippet").get_attribute("innerHTML")
            review_description = review_description.replace + more_description
        except:
            pass
        
        try:
            driver.find_elements_by_class_name("review-container")[i].find_element_by_class_name("ulBlueLinks").click()
        except:
            pass

        # ----------------------------- Date of visit to restaurant ----------------------------- #
        review_date_of_visit = date_of_visit_list[i]
        
        # ----------------------------- Image uploaded ----------------------------- #
        try:
            review_img_src = [element.find_element_by_tag_name("img").get_attribute("src") 
                for element in review_container[i].find_elements_by_class_name("imgWrap")]
        except:
            review_img_src = []

        df_temp = pd.DataFrame({
            "rating": [review_rating],
            "date": [review_date],
            "title": [review_title],
            "description": [review_description],
            "date_of_visit": [review_date_of_visit],
            "img_src": [review_img_src],
            "url": [restaurant_url]
        })

        df = df.append(df_temp)

    return df

# to store scraped dataset
df_scraped = pd.DataFrame()

# index of df_links to begin scrape
start = 0

# number of times to run the codes below
loop = 2

for i in range(1,loop):
    driver = webdriver.Chrome()
    driver.maximize_window()

    index_start = start + ((i-1) * 10)
    index_end = start + (i*10) - 1
    
    df_scraped = pd.DataFrame()
    for idx, row in df_links.loc[index_start:index_end,:].iterrows():
        driver.get(row["review_url"])
        sleep(10)        # shorter sleep time faster but easier to break from StaleElementReferenceException
        attempts = 0
        done = False
        while (attempts < 5) and (not done):
            try:
                df_scraped = scrape_page_review(df_scraped, row["restaurant_url"])
                print(f"Done {idx} of {index_end}")
                done = True
            except:
                attempts += 1
            
    # scraped file goes into review directory
    df_scraped.to_csv(f"review/reviews_from_{index_start}_to_{idx}.csv", encoding="utf-8-sig", index=False)
    driver.quit()

df_scraped.reset_index(drop=True, inplace=True)
df_scraped
