==> 04_scrape_shopping_mall_address.ipynb <==
# Finding the Restaurants that are within Shopping Malls
Notebook used to extract Shopping Mall data, and check if Restaurants exist in Shopping Malls

# conda install -c anaconda beautifulsoup4
from bs4 import BeautifulSoup, SoupStrainer
import requests
import re

import pandas as pd
from fuzzywuzzy import fuzz
import numpy as np

from selenium import webdriver      # conda install -c conda-forge selenium
from time import sleep
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.keys import Keys

import warnings
warnings.filterwarnings("ignore")

## 1. Web Scraping the list of Shopping Mall names from Wikipedia (using BeautifulSoup)

url = 'https://en.wikipedia.org/wiki/List_of_shopping_malls_in_Singapore'
resp = requests.get(url)
list_of_mall_names = SoupStrainer('div', class_='div-col')
soup = BeautifulSoup(resp.text, 'html.parser', parse_only=list_of_mall_names)

list_of_scraped_names = []

for li_tag in soup.find_all('li'):
    li_tag_text = li_tag.text
    extra_bracket_match = re.search('\[\d*\d]', li_tag_text)
    if extra_bracket_match:
        extra_bracket = extra_bracket_match.group()
        li_tag_text = li_tag_text.replace(extra_bracket, '').strip()
    list_of_scraped_names.append(li_tag_text)

df_malls = pd.DataFrame({'mall_names': list_of_scraped_names})
df_malls.drop_duplicates(inplace=True)

df_malls

## 2. Geocoding for Shopping Mall addresses

driver = webdriver.Chrome()
driver.get("https://www.google.com/maps")
dct_info = {}

for idx, row in df_malls.iterrows():
    mall = row['mall_names']
    input_element = driver.find_element_by_xpath('//*[@id="searchboxinput"]')
    input_element.clear()
    input_element.send_keys(f"{mall} location")
    input_element.send_keys(Keys.ENTER)
    sleep(5)
    try:
        first_reco = driver.find_element_by_class_name("a4gq8e-aVTXAb-haAclf-jRmmHf-hSRGPd")
        first_reco.click()
        sleep(5)
        print(mall)
    except:
        pass

    address_elements = driver.find_elements_by_class_name("QSFF4-text")
    if len(address_elements) > 0:
        address = address_elements[0].text
    else:
        address = ""

    url_current = driver.current_url
    match = re.search('@(-?\d+\.?\d*),(-?\d+\.?\d*)', url_current)
    lat = round(float(match.group(1).replace('@','').strip()), 6)
    lon = round(float(match.group(2)), 6)
    dct_info[mall] = [lat, lon, address]

dct_info['Marina Bay Sands'] = [1.2840501, 103.858707, '10 Bayfront Ave, Singapore 018956']
dct_info['The South Beach'] = [1.2957395, 103.8544249, '38 Beach Rd, Singapore 189767']
dct_info['Northshore Plaza'] = [1.4168672, 103.8996155, '407 Northshore Dr, Singapore 820407']
dct_info['Bedok Point'] = [1.3249305, 103.9301924, 'SG, New Upper Changi Rd, 799, 467351']
dct_info['Jurong Point'] = [1.3397443, 103.704541, '1 Jurong West Central 2, Singapore 648886']
dct_info['Anchorpoint'] = [1.2887706, 103.8028875, '370 Alexandra Rd, Singapore 159953']
dct_info['Pioneer Mall'] = [1.3418699, 103.6951971, '638 Jurong West Street 61, Singapore 640638']
dct_info["People's Park Complex"] = [1.2841977, 103.8403601, '1 Park Rd, Singapore 059108']
dct_info['Orchard Plaza'] = [1.301077, 103.8387947, '150 Orchard Rd, Orchard Plaza, Singapore 238841']
dct_info['Tekka Centre'] = [1.306256, 103.848471, '665 Buffalo Road, Singapore 210665']

df_malls['address'] = df_malls['mall_names'].apply(lambda x: dct_info[x][0])
df_malls['latitude'] = df_malls['mall_names'].apply(lambda x: dct_info[x][1])
df_malls['longitude'] = df_malls['mall_names'].apply(lambda x: dct_info[x][2])

df_malls.to_csv('data/singapore_mall_locations.csv', encoding='utf-8-sig', index=False)

## 3. Use Regular Expressions to identify Restaurants in Malls
Based on address and postal code

df_malls = pd.read_csv('data/singapore_mall_locations.csv')

df_malls

# extract postal code
df_malls['postal_code'] = df_malls['address'].apply(
    lambda x: re.search('\d{3}\s*\d{3}', x).group() if re.search('\d{3}\s*\d{3}', x) else np.nan
)

def split_sticked_postal_code(x):
    match = re.search('(\w*[\d{3}])(\d{3}\w*)', x)
    if match:
        first = match.groups()[0]
        second = match.groups()[1]
        new = first + ' ' + second
        x = x.replace(first+second, new)
        return x
    return x

df_restaurant = pd.read_csv("data/full_restaurant_rating_data.csv")

### Check if Mall Name Exist in Restaurant

df_restaurant['cleaned_address'] = df_restaurant['address'].str.lower()
df_restaurant['cleaned_address'] = df_restaurant['cleaned_address'].apply(split_sticked_postal_code)

df_malls['cleaned_mall_names'] = df_malls['mall_names'].str.lower()
df_malls['postal_code'] = df_malls['postal_code'].apply(split_sticked_postal_code)

def create_dict_from_df(df, col_key, col_value):
    df = df[[col_key, col_value]].drop_duplicates()
    return df.set_index([col_key])[col_value].to_dict()

dct_postal_code_mall = create_dict_from_df(df_malls, 'postal_code', 'mall_names')

def matches_shopping_mall(mall, restaurant):
    malls_to_fully_match = ['nex', 'imm', 'duo', 'jurong point', 'singpost centre']
    if mall in ['nex', 'imm', 'duo']:
        ### for these mall names, require them to be an individual word (not part of a bigger word like 'next' containing 'nex')
        tokens = restaurant.split(' ')
        return any([k == mall for k in tokens])
    elif mall in ['jurong point', 'singpost centre'] or re.search('\d+', mall):
        # mall names that is harder to be identified by FuzzyWuzzy
        return mall in restaurant
    else:
        # FuzzyWuzzy partio ratio returns String Matching scores (account for different variations of a mall name)
        return fuzz.partial_ratio(mall, restaurant) > 90

for idx, row in df_restaurant.iterrows():
    x = row['cleaned_address']
    name_match = [k for k in df_malls['cleaned_mall_names'] if matches_shopping_mall(k, x)]
    postal_code_match = [dct_postal_code_mall[k] for k in df_malls['postal_code'] if k in x]
    
    if any(name_match) or any(postal_code_match):
        df_restaurant.loc[idx, 'is_in_mall'] = 1
    else:
        df_restaurant.loc[idx, 'is_in_mall'] = 0

    if len(name_match) > 0:
        df_restaurant.loc[idx, 'mall'] = name_match[0].title()
        continue
    if len(postal_code_match) > 0:
        df_restaurant.loc[idx, 'mall'] = postal_code_match[0].title()

## 4. Inspect New Features Generated

df_subset = df_restaurant[['url', 'is_in_mall', 'mall']]

df_subset

df_subset.to_csv('data/restaurant_is_in_mall.csv', index=False, encoding='utf-8-sig')
