==> 05_scrape_train_stn_walk_distance.ipynb <==
# Finding the Walking Duration of each Restaurant from the nearest Station Exit
Notebook used to extract walking distance between restaurant and nearest train station exit

import pandas as pd
import numpy as np
import re

import geopandas as gpd #conda install geopandas
# conda install -c conda-forge nbformat

import plotly.express as px #conda install -c plotly plotly_express
from plotly.express import choropleth_mapbox    
from shapely.geometry import shape, Point
import openrouteservice

import warnings
warnings.filterwarnings("ignore")

## 1. Basic Setup, Loading Datasets

# ---------------- Loading in Restaurant dataset ---------------- #
df_restaurant = pd.read_csv("data/full_restaurant_rating_data.csv")

# ---------------- Converting Dataset to GeoDataFrame ---------------- #
gdf_restaurant = gpd.GeoDataFrame(df_restaurant, geometry= gpd.points_from_xy(df_restaurant['longitude'], df_restaurant['latitude']), crs='epsg:4326')

# ---------------- Loading in Train Station Exit Locations ---------------- #
### Dataset does not have the train station name, so need to use Train Station Locations dataset to identify
gdf_staion_exit = gpd.read_file('data/train_station_exit.geojson', crs='epsg:4326')

# ---------------- Loading in Train Station Locations ---------------- #
df_station = pd.read_csv('data/train_station_location.csv')
gdf_station = gpd.GeoDataFrame(df_station, geometry= gpd.points_from_xy(df_station['Longitude'], df_station['Latitude']), crs='epsg:4326')

gdf_staion_exit

gdf_station

## 2. Assign Train station of each station exits

gdf_staion_exit['key'] = 1
gdf_station['key'] = 1

# dataframe where each station exit is mapped to each station
merged_station_exit = pd.merge(gdf_staion_exit, gdf_station, on='key').drop('key', 1)

# GeoSeries of all the Coordinates of Station Exits
geoseries_exit = gpd.GeoSeries(merged_station_exit['geometry_x'], crs='EPSG:4326').to_crs('EPSG:24500')
# GeoSeries of all the Coordinates of Station
geoseries_station = gpd.GeoSeries(merged_station_exit['geometry_y'], crs='EPSG:4326').to_crs('EPSG:24500')

# Computing Distance (in meters) of each exit to each station
merged_station_exit['distance'] = geoseries_exit.distance(geoseries_station)

# Obtaining the nearest station for each exit
df_nearest_station_to_each_exit = merged_station_exit.sort_values(['Name', 'distance'], ascending=True).groupby('Name').head(1)

# Dropping unnecessary columns
df_nearest_station_to_each_exit.drop(columns=['Description', 'X', 'Y', 'Latitude', 'Longitude', 'geometry_y'], inplace=True)
df_nearest_station_to_each_exit.rename(columns={'geometry_x': 'geometry'}, inplace=True)

# Converting to GeoDataFrame
df_nearest_station_to_each_exit['geometry'] = gpd.GeoSeries(df_nearest_station_to_each_exit['geometry'], crs='EPSG:4326')
gdf_nearest_station_to_each_exit = gpd.GeoDataFrame(df_nearest_station_to_each_exit, geometry= df_nearest_station_to_each_exit.geometry, crs='epsg:4326')

## 3. Find nearest train station exit for each restaurant

gdf_nearest_station_to_each_exit['key'] = 1
gdf_restaurant['key'] = 1

# dataframe where each station exit is mapped to each station
merged_exit_restaurant = pd.merge(gdf_nearest_station_to_each_exit, gdf_restaurant, on='key').drop('key', 1)

# GeoSeries of all the Coordinates of Station Exits
geoseries_exit = gpd.GeoSeries(merged_exit_restaurant['geometry_x'], crs='EPSG:4326').to_crs('EPSG:24500')
# GeoSeries of all the Coordinates of Restaurants
geoseries_restaurant = gpd.GeoSeries(merged_exit_restaurant['geometry_y'], crs='EPSG:4326').to_crs('EPSG:24500')

# Computing Distance (in meters) of each restaurant to each exit
merged_exit_restaurant['distance'] = geoseries_exit.distance(geoseries_restaurant)

# Obtaining the nearest exit for each restaurant
df_restaurant_nearest_exit = merged_exit_restaurant.sort_values(['url', 'distance'], ascending=True).groupby('url').head(1)

# Dropping unnecessary columns
df_restaurant_nearest_exit = df_restaurant_nearest_exit[['Name', 'geometry_x', 'STN_NAME', 'STN_NO', 
                    'COLOR', 'distance', 'url', 'geometry_y']].copy()

df_restaurant_nearest_exit.rename(columns={
    'Name': 'exit_name',
    'STN_NAME': 'station_name',
    'STN_NO': 'station_no',
    'COLOR': 'station_line_color',
    'geometry_x': 'exit_geometry',
    'geometry_y': 'restaurant_geometry'
}, inplace=True)

# Converting to GeoDataFrame
df_restaurant_nearest_exit.reset_index(drop=True, inplace=True)

# df_restaurant_nearest_exit.to_csv('data/restaurant_nearest_station_exit.csv', index=False, encoding='utf-8-sig')

df_restaurant_nearest_exit

## 4. Finding Walking Distance and Duration of Train Exit from Restaurant

#### Reading in Nearest Restaurant Exit Dataset

df_restaurant_nearest_exit = pd.read_csv('data/restaurant_nearest_station_exit.csv')

def extract_longitude(x):
    lon_lat_list = re.findall('\d+\.\d+', x)
    if len(lon_lat_list) > 1:
        return float(lon_lat_list[0])
    return np.nan

def extract_latitude(x):
    lon_lat_list = re.findall('\d+\.\d+', x)
    if len(lon_lat_list) > 1:
        return float(lon_lat_list[1])
    return np.nan

# geometry columns read in as string by read_csv
df_restaurant_nearest_exit['exit_lon'] = df_restaurant_nearest_exit['exit_geometry'].apply(extract_longitude)
df_restaurant_nearest_exit['exit_lat'] = df_restaurant_nearest_exit['exit_geometry'].apply(extract_latitude)
df_restaurant_nearest_exit['restaurant_lon'] = df_restaurant_nearest_exit['restaurant_geometry'].apply(extract_longitude)
df_restaurant_nearest_exit['restaurant_lat'] = df_restaurant_nearest_exit['restaurant_geometry'].apply(extract_latitude)

#### Use OpenRouteServices to extract walking distance and duration

# ORS API token
try:
    from tokens import ors_token
except:
    ors_token = '############### USE OWN ORS TOKEN ###############'

# -------- Function to use Openrouteservice API to obtain walking distance and duration --- #
def find_travelling_route(start_pt_lon, start_pt_lat, end_pt_lon, end_pt_lat, ors_token):
    client = openrouteservice.Client(key=ors_token)
    coord = ((start_pt_lon, start_pt_lat), (end_pt_lon, end_pt_lat))
    route = client.directions(coord, profile='foot-walking')
    return route

# Note that ORS has a daily limit of 2000, so only 2000 routes can be retrieved
for idx, row in df_restaurant_nearest_exit.iterrows():
    x_lon = row['exit_lon']
    x_lat = row['exit_lat']
    y_lon = row['restaurant_lon']
    y_lat = row['restaurant_lat']

    # empty latitude or longitude
    if pd.isna(y_lon) or pd.isna(y_lat) or pd.isna(x_lat) or pd.isna(x_lon):
        continue

    try:
        route = find_travelling_route(x_lon, x_lat, y_lon, y_lat, ors_token)
        if 'distance' in route['routes'][0]['summary']:
            df_restaurant_nearest_exit.loc[idx, 'walking_distance'] = route['routes'][0]['summary']['distance']
            df_restaurant_nearest_exit.loc[idx, 'walking_duration'] = route['routes'][0]['summary']['duration']
    except:
        df_restaurant_nearest_exit.loc[idx, 'walking_distance'] = np.nan
        df_restaurant_nearest_exit.loc[idx, 'walking_duration'] = np.nan

df_restaurant_nearest_exit
